+++ components/OsIcon/index.ts
@@ -1,2 +1,1 @@
-export { OsIcon, type Os } from './OsIcon';
-export { OsIcon as default } from './OsIcon';
+export { type Os, OsIcon, OsIcon as default } from './OsIcon';